% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carf.R
\name{carf_check_dig}
\alias{carf_check_dig}
\title{Validate check digits for Brazilian lawsuits identification
number}
\usage{
carf_check_dig(id)
}
\arguments{
\item{id}{String containing the complete lawsuit number}
}
\value{
Whether or not the check digit is well calculated
}
\description{
Verifies if a check digit is correct
}
\examples{
{
carf_check_dig("10120.008427/2003-02")
carf_check_dig(c("10120008427200302", "10766.000511/96-12"))
}
}
