% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj.R
\name{check_dig}
\alias{check_dig}
\title{Validate check digits for Brazilian lawsuits identification
number}
\usage{
check_dig(num)
}
\arguments{
\item{num}{String containing the complete lawsuit number}
}
\value{
Whether or not the check digit is well calculated
}
\description{
Verifies if a check digit is correct
}
\examples{
{
check_dig("0005268-75.2013.8.26.0100")

\dontrun{
check_dig("0005268-75.2013.8.26.100", build = TRUE)
}
}

}
