% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj.R
\name{check_dig_vet}
\alias{check_dig_vet}
\title{Validate check digits for Brazilian lawsuits identification
number on vectorial scale.}
\usage{
check_dig_vet(num)
}
\arguments{
\item{num}{A vector containing strings with the complete lawsuit number}
}
\value{
Whether or not the check digit is well calculated
}
\description{
Verifies if a check digit is correct
}
\examples{
{
check_dig_vet(c("0005268-75.2013.8.26.0100", "0004122-85.2010.6.16.0100"))

}

}
