% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chrome_to_body}
\alias{chrome_to_body}
\title{Convert Chrome's Query String Parameters to a list}
\usage{
chrome_to_body(x)
}
\arguments{
\item{x}{A string with Chrome's Query String Parameters}
}
\description{
To use this function, simply copy the Query String
Parameters returned by Chrome when analyzing the network flow of
a web page. Paste these QSPs into an R string with double quotes
(as you would to create any string) and pass it to
\code{chrome_to_body()}; the function will print to the console a
formatted command that creates a list with the QSPs. This list
works perfectly with \code{\link[httr:GET]{httr::GET()}} and \code{\link[httr:POST]{httr::POST()}} so that
you can easily reproduce a website's behavior.
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}
}
