% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dvec}
\alias{dvec}
\title{Vetorizando scrapers}
\usage{
dvec(fun, itens, ..., verbose = TRUE, p = 0.05)
}
\arguments{
\item{fun}{funcao a ser aplicada em cada arquivo.}

\item{itens}{character vector dos caminhos de arquivos a serem transformados.}

\item{...}{outros parametros a serem passados para \code{fun}}

\item{verbose}{se \code{TRUE} (default), mostra o item com probabilidade p.}

\item{p}{probabilidade de imprimir mensagem.}
}
\description{
Vetoriza um scraper (funcao) para um vetor de itens
}

