% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_unicode.R
\name{escape_unicode}
\alias{escape_unicode}
\title{Escape accented characters in a document}
\usage{
escape_unicode()
}
\description{
This function is used by the "Escape Unicode" add-in
and removes all accented characters from the current file, replacing
them by their equivalent Unicode-escaped values.
}
