% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_sans_ext}
\alias{file_sans_ext}
\title{Extract file name without extension}
\usage{
file_sans_ext(x)
}
\arguments{
\item{x}{Character vector of file paths}
}
\description{
Extract file name without extension
}
