% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj.R
\name{gather_subjects}
\alias{gather_subjects}
\title{Gather subjects from esaj::cjsg_table("subjects")}
\usage{
gather_subjects(subjects)
}
\arguments{
\item{subjects}{Table returned by \code{esaj::cjsg_table("subjects")}}
}
\description{
Once you run \code{esaj::cjsg_table("subjects")}, you can
use this function to gather the subjects automatically. Download
\code{esaj} by running \code{devtools::install_github("courtsbr/esaj")}.
}
