% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{github_template}
\alias{github_template}
\title{Criar repositório e template de R-package}
\usage{
github_template(repo, org = "", path = ".", descr = NULL)
}
\arguments{
\item{repo}{nome do repo.}

\item{org}{nome da organização.}

\item{path}{caminho do repo.}

\item{descr}{descrição do pacote. Se nulo, tentará pegar de \code{Sys.getenv('REPO_DESC')}.}
}
\description{
Só funciona se tiver um \code{Sys.getenv("GITHUB_PAT")}.
}

