% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lsos}
\alias{lsos}
\title{Lista objetos.}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{outros parametros passados para \code{.ls.objects}.}

\item{n}{apenas os n com maior tamanho.}
}
\description{
Lista objetos.
}

