% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pvec}
\alias{pvec}
\title{Verbose, parallel, and safe map-like}
\usage{
pvec(.x, .f, ..., .cores = get_cores(), .progress = TRUE,
  .flatten = FALSE)
}
\arguments{
\item{.x}{A list or atomic vector}

\item{.f}{A function, formula, or atomic vector (see \code{\link[purrr:map]{purrr::map()}})}

\item{...}{Other parameters passed on to \code{\link[furrr:future_map]{furrr::future_map()}}}

\item{.cores}{Number of cores to use when multiprocessing}

\item{.progress}{Whether or not to display progress}

\item{.flatten}{Whether or not to return only the output of the
function (replaces errors with \code{NA})}
}
\description{
Using the same argument notation as \code{\link[purrr:map]{purrr::map()}}, this function
iterates over a list of inputs \code{.x}, applying \code{.f} to each element. It
returns a tibble with the input, whether the function returned an error
and the output.
}
