% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reais}
\alias{reais}
\title{Convert brazilian currency values (text) to numeric}
\usage{
reais(x)
}
\arguments{
\item{x}{A currency vector. Ex: c("R$ 10.000,00", "R$ 123,00")}
}
\description{
Convert brazilian currency values (text) to numeric
}
