% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_accent}
\alias{rm_accent}
\title{Remove accents.}
\usage{
rm_accent(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
A version of x without non-ASCII characters.
}
\description{
Remove accents from strings converting them to ASCII.
}

