% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_accent_from_names.R
\name{rm_accent_from_names}
\alias{rm_accent_from_names}
\title{Remove accentuation from column names (DEPRECATED)}
\usage{
rm_accent_from_names(dat)
}
\arguments{
\item{dat}{A dataset}
}
\value{
A version of dat without non-ASCII characters.
}
\description{
Remove accented characters from column names converting them to ASCII.
}
