% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cnj.R
\name{sample_cnj}
\alias{sample_cnj}
\title{Sample brazillian process codes.}
\usage{
sample_cnj(n, foros, anos, orgao, tr, first_dig = "0", sample_pars = T,
  return_df = T)
}
\arguments{
\item{n}{a non negative integer giving the number of process codes to sample.}

\item{foros}{a string scalar (or vector) with 4 characters. Identifies the place
where the sampled processes need to be distributed.}

\item{anos}{a string scalar (or vector) with 4 characters. Identifies the 
distribution years of the sampled codes.}

\item{orgao}{a string scalar (or vector) with 1 character. Identifies the jurisdiction
of the sampled codes.}

\item{tr}{a string scalar (or vector) with 2 character. Identifies the court of 
the sampled codes.}

\item{first_dig}{the first digit of the process code. It's usually "0" or "1". 
"0" as detault. The first digit will be sampled if first_dig = "".}

\item{sample_pars}{Logical scalar.
Does the parameters define the characteristics of the codes or should be sampled
as well?}

\item{return_df}{Logical scalar. Should the function return a df? If FALSE the
function returns a vector.}
}
\value{
A data_frame or a vector contaning a random sample of processes of size n.
}
\description{
\code{sample_cnj} returns a data_frame containing a random sample of
process codes sampled according to some regional and jurisdictional 
parameters. The actual implementation supports both vector or scalar parameters,
depending wheter or not one whants to sample from a determined scope of
process codes or whants to define every single set of parameters.
}

