% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj.R
\name{separate_cnj}
\alias{separate_cnj}
\title{Separate a lawsuit ID column into its parts}
\usage{
separate_cnj(data, col, ...)
}
\arguments{
\item{data}{A data frame}

\item{col}{Column name or position (see \code{\link[tidyr:separate]{tidyr::separate()}})}

\item{...}{Other arguments passed on to \code{\link[tidyr:separate]{tidyr::separate()}}}
}
\description{
Wrapper around \code{\link[tidyr:separate]{tidyr::separate()}} that splits a column
with lawsuit IDs into 6 columns with its parts (see \code{\link[=extract_parts]{extract_parts()}}).
Note that the IDs must be built (see \code{\link[=build_id]{build_id()}}).
}
