% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tit.R
\name{tit_calc_dig}
\alias{tit_calc_dig}
\title{Calculate check digit for TIT}
\usage{
tit_calc_dig(id, build = FALSE)
}
\arguments{
\item{id}{Lawsuit number (including trailling zeros), excluding
the check digit.}

\item{build}{Whether or not the function return the complete number
(or only the check digits)?}
}
\value{
The check digits or the complete identification number
}
\description{
Returns the check digit of a TIT number or full number with
the check digit.
}
\examples{
{
tit_calc_dig("3107557", build = TRUE)
}

}
