% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tit.R
\name{tit_check_dig}
\alias{tit_check_dig}
\title{Validate check digits for Brazilian lawsuits identification
number}
\usage{
tit_check_dig(id)
}
\arguments{
\item{id}{String containing the complete lawsuit number}
}
\value{
Whether or not the check digit is well calculated
}
\description{
Verifies if a check digit is correct
}
\examples{
{
tit_check_dig("3107557-5")
tit_check_dig(c("30385040", "10008596-9"))
}
}
