% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_pipe}
\alias{use_pipe}
\title{Add pipe template}
\usage{
use_pipe(pkg = ".")
}
\arguments{
\item{pkg}{Package description (can be path or package name)}
}
\description{
Adds pipe template to package documentation.
}
