% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj.R
\name{verify_cnj}
\alias{verify_cnj}
\title{Validate Brazilian lawsuits identification number on vectors.}
\usage{
verify_cnj(cnj)
}
\arguments{
\item{cnj}{A vector containing strings with the complete lawsuit number}
}
\value{
Whether or not the check digit is well calculated
}
\description{
Verifies if a Brazilian lawsuit identification is a cnj number.
}
