% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_2p}
\alias{with_2p}
\title{Add paralellism and \code{\link[purrr:possibly]{purrr::possibly()}} to any mapping function}
\usage{
with_2p(..f)
}
\arguments{
\item{..f}{A function, formula, or atomic vector (see \code{\link[purrr:map]{purrr::map()}})}
}
\description{
This function wraps any function of the \code{\link[purrr:map]{purrr::map()}} family,
running everything in parallel and capturing side effects with \code{\link[purrr:possibly]{purrr::possibly()}}.
This is useful when only the side effects are desired as the returned values
will be either \code{TRUE} or \code{FALSE} (if that iteration did/didn't work).
}
