% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_parallelism}
\alias{with_parallelism}
\title{Add parallelism to any mapping function}
\usage{
with_parallelism(..f)
}
\arguments{
\item{..f}{A function, formula, or atomic vector (see \code{\link[purrr:map]{purrr::map()}})}
}
\description{
This function wraps any function of the \code{\link[purrr:map]{purrr::map()}} family,
parallelizing its iterations with \code{\link[parallel:mcmapply]{parallel::mcmapply()}}. Note that flattening
doesn't work, so \code{purrr::map_***()} becomes only \code{purrr::map()}.
}
