% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_progress}
\alias{with_progress}
\title{Add a progress bar to any mapping function}
\usage{
with_progress(..f)
}
\arguments{
\item{..f}{A function, formula, or atomic vector (see \code{\link[purrr:map]{purrr::map()}})}
}
\description{
This function wraps any function of the \code{\link[purrr:map]{purrr::map()}} family,
adding a progress bar to the iteration.
}
