% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_data}
\alias{write_data}
\title{Shortcut to write file to "data/" directory from a pipe}
\usage{
write_data(x, name, dir = "data/")
}
\arguments{
\item{x}{Object to write}

\item{name}{Name of the object (important when loading)}

\item{dir}{Directory where to save file}
}
\description{
Shortcut to write file to "data/" directory from a pipe
}
