% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj.R
\name{calc_dig}
\alias{calc_dig}
\title{Calculate digits for Brazilian lawsuit identification numbers}
\usage{
calc_dig(num, build = FALSE)
}
\arguments{
\item{num}{Ordered digits of the lawsuit number (including 0's) excluding
the check digit}

\item{build}{Whether or not the function return the complete lawsuit
number (or only the check digits)?}
}
\value{
The check digits or the complete identification number
}
\description{
Returns the check digit of a lawsuit numbers in the format
unified by the brazillian National Council of Justice.
}
\examples{
{
calc_dig("001040620018260004", build = TRUE)
calc_dig("001040620018260004", build = FALSE)
\dontrun{
calc_dig("00104062001826000", build = TRUE)
}
}

}
