% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dig.R
\name{check_dig}
\alias{check_dig}
\title{Validate check digits for brazilian lawsuits identification number}
\usage{
check_dig(num)
}
\arguments{
\item{num}{String scalar containing the complete docket number.}
}
\value{
Logical scalar indicating whether or not the check digit is well 
  calculated.
}
\description{
\code{check_dig} verifies if a check digit is correct.
}
\examples{

check_dig("0005268-75.2013.8.26.0100")

\dontrun{
check_dig("0005268-75.2013.8.26.100", build = TRUE)
}
}

