% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dvec}
\alias{dvec}
\title{Vectorize functions}
\usage{
dvec(fun, itens, ..., verbose = TRUE, p = 0.05)
}
\arguments{
\item{fun}{function to be iterated.}

\item{itens}{character vector of inputs.}

\item{...}{outros other parameters for \code{fun}}

\item{verbose}{should dvec print the current "item"?
if \code{TRUE} (default) shows a message with probability p.}

\item{p}{probability of printing a message. Only meaningful when verbose is 
\code{TRUE}.}
}
\description{
Iterate a function and wrap a dplyr::failtwith around it.
}

