% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
lsos(pos = 1, pattern, order.by = "Size", decreasing = TRUE,
  head = TRUE, n = 10)
}
\arguments{
\item{pos}{Where to look for the object (see "Details" in \code{\link[base:get]{base::get()}}'s
documentation)}

\item{pattern}{An optional regular expression to match names (\code{\link[utils:glob2rx]{utils::glob2rx()}}
can be used to convert wildcard patterns to regular expressions)}

\item{order.by}{Sort by \code{"Size"} (default), \code{"Type"}, \code{"Rows"} or \code{"Columns"}}

\item{decreasing}{Should the sorting be decreasing?}

\item{head}{Should \code{\link[utils:head]{utils::head()}} function be used for printing?}

\item{n}{How many lines \code{\link[utils:head]{utils::head()}} function should show?}
}
\description{
Elegantly list objects in a R session.
}
\references{
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
