% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_accent}
\alias{rm_accent}
\title{Remove accentuation}
\usage{
rm_accent(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
A version of \code{x} without non-ASCII characters
}
\description{
Remove accented characters from strings converting them to
ASCII.
}
